Added Naive Single end stat algorithm
This algorithm to determine the current state will not handle items that
can exist in multiple states correctly. For example, if an item can be
acted on by multiple parties A & B and be transitioned into multiple
states, then the algorithm would fail to take into account the state and
actions themselves.