Add test for multi-end state workflow
A bigger question is how one would be able to decide if you could
arbitrarily add in an action. But I suppose that's another detail to be
handled somewhere else. I mean, it's basically saying that you can
always start a null state to X state. Which would make sense anyway
somewhat.