Add tests for new methods and for multi-end actions
The multi-end action test might be wrong right now, but it does show an
issue in the current algorithm. Which is that we dont short circuit the
stack pushing once we find one. which means if we have a workflow like
this:

    s1 -- s2 -- s3 -- s4
           | -- s5

And the log has something like:

	∅ → s1
	s1 → s2
    s3 → s2 (loop)
    ∅ → s1 (new seq)
    s1 → s2 (new seq moves to s2)

We'll have, in the sequences array:

	Stack [ s3 → s2 ] (top)
	Stack [ s1 → s2 ] (top)

Which means that the possible actions will be only for s2. Which is
technically correctly, but we lose the seperate threads going on and a
log of

	s2 → s3

will cause both of the stacks to increment and not just one of them. Is
that what makes sense to do? I think it doesn't. As once you have
multiple paths going on it needs to complete both of them. We don't have
a good strategy for choosing _which_ stack to increment though. So I
need to think on that a little bit.