Handle multi-end state possible actions better
Handling the stack of actions independently from each other now so that
we only apply the next logentry _once_ to the first stack that can take
it. Without actually having some type of key to tell if an set of log
entries should be grouped together or not there's not really a better
option. This should allow for something like the following to happen:

	s0 -> s1
	s1 -> s2
	s0 -> s3
	s3 -> s2
	-- then --
	s2 -> s4 (take action, but still have option to do next)
	s2 -> s5

It makes sense either way to allow a path, once it's split to continue
being two independent paths I think.